+++ gno.land/cmd/tm2txsync/import.go
@@ -16,7 +16,7 @@ import (
 
 	_ "github.com/gnolang/gno/tm2/pkg/sdk/auth" // XXX better way?
 	_ "github.com/gnolang/gno/tm2/pkg/sdk/bank"
-	_ "github.com/gnolang/gno/gnovm/pkg/sdk/vm"
+	_ "github.com/gnolang/gno/gno.land/pkg/sdk/vm"
 )
 
 type importCfg struct {
