+++ gno.land/pkg/sdk/vm/invariants.go
