+++ gno.land/pkg/sdk/vm/package.go
@@ -6,7 +6,7 @@ import (
 )
 
 var Package = amino.RegisterPackage(amino.NewPackage(
-	"github.com/gnolang/gno/gnovm/pkg/sdk/vm",
+	"github.com/gnolang/gno/gno.land/pkg/sdk/vm",
 	"vm",
 	amino.GetCallersDirname(),
 ).WithDependencies(
