+++ gno.land/pkg/sdk/vm/vm.proto
@@ -1,7 +1,7 @@
 syntax = "proto3";
 package vm;
 
-option go_package = "github.com/gnolang/gno/gnovm/pkg/sdk/vm/pb";
+option go_package = "github.com/gnolang/gno/gno.land/pkg/sdk/vm/pb";
 
 // imports
 import "github.com/gnolang/gno/tm2/pkg/std/std.proto";
