+++ misc/docker-integration/integration_test.go
@@ -14,7 +14,7 @@ import (
 
 	"github.com/gnolang/gno/gno.land/pkg/gnoland"
 	"github.com/gnolang/gno/tm2/pkg/amino"
-	"github.com/gnolang/gno/gnovm/pkg/sdk/vm"
+	"github.com/gnolang/gno/gno.land/pkg/sdk/vm"
 	"github.com/gnolang/gno/tm2/pkg/std"
 	"github.com/stretchr/testify/require"
 	"gopkg.in/yaml.v3"
