+++ tm2/pkg/crypto/keys/client/addpkg.go
@@ -7,12 +7,12 @@ import (
 	"flag"
 	"fmt"
 
+	"github.com/gnolang/gno/gno.land/pkg/sdk/vm"
 	gno "github.com/gnolang/gno/gnovm/pkg/gnolang"
 	"github.com/gnolang/gno/tm2/pkg/amino"
 	"github.com/gnolang/gno/tm2/pkg/commands"
 	"github.com/gnolang/gno/tm2/pkg/crypto/keys"
 	"github.com/gnolang/gno/tm2/pkg/errors"
-	"github.com/gnolang/gno/gnovm/pkg/sdk/vm"
 	"github.com/gnolang/gno/tm2/pkg/std"
 )
